PSAT/FSAT mod notes with SummaryNote object in summary.component
