+++ src/app/fsat/fsat-report/results-summary/results-summary.component.ts
@@ -3,7 +3,7 @@ import { Assessment } from '../../../shared/models/assessment';
 import { Settings } from '../../../shared/models/settings';
 import { FSAT } from '../../../shared/models/fans';
 import { CompareService } from '../../compare.service';
-import { FsatReportRollupService, SummaryNote } from '../../../report-rollup/fsat-report-rollup.service';
+import { FsatReportRollupService } from '../../../report-rollup/fsat-report-rollup.service';
 
 @Component({
   selector: 'app-results-summary',
@@ -43,7 +43,7 @@ export class ResultsSummaryComponent implements OnInit {
     }
 
     if(this.fsat.modifications){
-      this.notes = this.fsatReportRollupService.buildSummaryNotes(this.fsat);
+      this.notes = this.buildSummaryNotes(this.fsat);
     }
 
   }
@@ -72,5 +72,47 @@ export class ResultsSummaryComponent implements OnInit {
   useModification() {
     this.fsatReportRollupService.updateSelectedFsats({ assessment: this.assessment, settings: this.settings }, this.selectedModificationIndex);
   }
+  
+  buildSummaryNotes(fsat: FSAT): Array<SummaryNote>{
+    let tmpNotesArr: Array<SummaryNote> = new Array<SummaryNote>();
+    fsat.modifications.forEach(mod =>{
+      if(mod.fsat.notes){
+        if(mod.fsat.notes.fluidNotes){
+          let note = this.buildNoteObject(mod.fsat.name, 'Fluid', mod.fsat.notes.fluidNotes);
+          tmpNotesArr.push(note);
+        }
+        if(mod.fsat.notes.fanSetupNotes){
+          let note = this.buildNoteObject(mod.fsat.name, 'Fan', mod.fsat.notes.fanSetupNotes);
+          tmpNotesArr.push(note);
+        }
+        if(mod.fsat.notes.fanMotorNotes){
+          let note = this.buildNoteObject(mod.fsat.name, 'Motor', mod.fsat.notes.fanMotorNotes);
+          tmpNotesArr.push(note);
+        }
+        if(mod.fsat.notes.fieldDataNotes){
+          let note = this.buildNoteObject(mod.fsat.name, 'Field Data', mod.fsat.notes.fieldDataNotes);
+          tmpNotesArr.push(note);
+        }
 
+      }
+
+    });
+    return tmpNotesArr;
+  }
+
+  buildNoteObject(modName: string, modMade: string, modNote: string): SummaryNote {
+    let summaryNote: SummaryNote = {
+      modName: modName,
+      modMade: modMade,
+      modNote: modNote
+    };
+    return summaryNote;
+  }
+
+}
+
+export interface SummaryNote {
+  modName: string;
+  modMade: string;
+  modNote: string;
 }
