+++ src/app/psat/psat-report/output-summary/output-summary.component.ts
@@ -3,7 +3,7 @@ import { PSAT } from '../../../shared/models/psat';
 import { Settings } from '../../../shared/models/settings';
 import { Assessment } from '../../../shared/models/assessment';
 import { CompareService } from '../../compare.service';
-import { PsatReportRollupService, SummaryNote } from '../../../report-rollup/psat-report-rollup.service';
+import { PsatReportRollupService } from '../../../report-rollup/psat-report-rollup.service';
 
 @Component({
   selector: 'app-output-summary',
@@ -42,7 +42,7 @@ export class OutputSummaryComponent implements OnInit {
     }
 
     if(this.psat.modifications){
-      this.notes = this.psatReportRollupService.buildSummaryNotes(this.psat);
+      this.notes = this.buildSummaryNotes(this.psat);
     }
 
   }
@@ -88,4 +88,42 @@ export class OutputSummaryComponent implements OnInit {
     }
     return result;
   }
+
+  buildSummaryNotes(psat: PSAT): Array<SummaryNote>{
+    let tmpNotesArr: Array<SummaryNote> = new Array<SummaryNote>();
+    psat.modifications.forEach(mod =>{
+      if(mod.notes){
+        if(mod.notes.pumpFluidNotes){
+          let note = this.buildNoteObject(mod.psat.name, 'Pump and Fluid', mod.notes.pumpFluidNotes);
+          tmpNotesArr.push(note);
+        }
+        if(mod.notes.motorNotes){
+          let note = this.buildNoteObject(mod.psat.name, 'Motor', mod.notes.motorNotes);
+          tmpNotesArr.push(note);
+        }
+        if(mod.notes.fieldDataNotes){
+          let note = this.buildNoteObject(mod.psat.name, 'Field Data', mod.notes.fieldDataNotes);
+          tmpNotesArr.push(note);
+        }
+      }
+    });
+    return tmpNotesArr;
+  }
+
+  buildNoteObject(modName: string, modMade: string, modNote: string): SummaryNote {
+    let summaryNote: SummaryNote = {
+      modName: modName,
+      modMade: modMade,
+      modNote: modNote
+    };
+    return summaryNote;
+  }
+
 }
+
+export interface SummaryNote {
+  modName: string;
+  modMade: string;
+  modNote: string;
+}
+
